Module Logs_async_reporter

val reporter : unit -> Logs.reporter

val json_reporter : unit -> Logs.reporter

val output_reporter : (bytes -> int -> int -> unit) -> Logs.reporter

val zstd_reporter : ?zstd:Zstandard.Streaming.Compression.t -> ?inbuf:Core.Bigbuffer.t -> ?outbuf:Core.Bigstring.t -> Async.Writer.t -> Logs.reporter * (unit -> unit)

val level_arg : Logs.level option Async.Command.Arg_type.t

Argument type to be used for use in Command params.

val set_level_via_param : ?arg_name:string -> ?doc:string -> Logs.src list -> unit Async.Command.Param.t

set_level_via_param src is a param that sets the level of srcs (or all srcs if src is None).

val set_color_via_param : ?arg_name:string -> ?doc:string -> unit -> unit Async.Command.Param.t

set_color_via_param src is a param that sets the level of srcs (or all srcs if src is None).